+++ unit_test/mpi/test_dmrg_sa_tto_n2_sto3g.cpp
@@ -42,7 +42,7 @@ class TestTTODMRGN2STO3GSA : public ::testing::Test {
     typedef double FP;
 
     template <typename S, typename FL>
-    void test_dmrg(const vector<S> &targets, const vector<double> &energies,
+    void test_dmrg(const vector<S> &targets, const vector<long double> &energies,
                    const shared_ptr<HamiltonianQC<S, FL>> &hamil,
                    const string &name, ubond_t bond_dim, uint16_t nroots,
                    int tto);
@@ -67,7 +67,7 @@ bool TestTTODMRGN2STO3GSA::_mpi = MPITest::okay();
 
 template <typename S, typename FL>
 void TestTTODMRGN2STO3GSA::test_dmrg(
-    const vector<S> &targets, const vector<double> &energies,
+    const vector<S> &targets, const vector<long double> &energies,
     const shared_ptr<HamiltonianQC<S, FL>> &hamil, const string &name,
     ubond_t bond_dim, uint16_t nroots, int tto) {
 
@@ -142,7 +142,7 @@ void TestTTODMRGN2STO3GSA::test_dmrg(
     dmrg->solve(tto, mps->center == 0, 0);
 
     me->dot = 1;
-    double energy = dmrg->solve(10, mps->center == 0, 1E-8);
+    long double energy = dmrg->solve(10, mps->center == 0, 1E-8);
 
     // deallocate persistent stack memory
     mps_info->deallocate();
@@ -192,7 +192,7 @@ TEST_F(TestTTODMRGN2STO3GSA, TestSU2) {
                 if (na - nb >= 0)
                     targets.push_back(SU2(na + nb, na - nb, i));
 
-    vector<double> energies = {
+    vector<long double> energies = {
         -107.654122447525, // < N=14 S=0 PG=0 >
         -107.356943001688, // < N=14 S=1 PG=2|3 >
         -107.356943001688, // < N=14 S=1 PG=2|3 >
@@ -238,7 +238,7 @@ TEST_F(TestTTODMRGN2STO3GSA, TestSZ) {
             for (int nb = ne - 1; nb <= ne + 1; nb++)
                 targets.push_back(SZ(na + nb, na - nb, i));
 
-    vector<double> energies = {
+    vector<long double> energies = {
         -107.654122447526, // < N=14 S=0 PG=0 >
         -107.356943001689, // < N=14 S=-1|0|1 PG=2|3 >
         -107.356943001688, // < N=14 S=-1|0|1 PG=2|3 >
