+++ unit_test/mpi/test_real_te_h10_sto6g.cpp
@@ -219,7 +219,7 @@ void TestRealTEH10STO6G::test_dmrg(
         make_shared<DMRG<S, FL, FL>>(me, ket_bdims, noises);
     dmrg->noise_type = NoiseTypes::ReducedPerturbative;
     dmrg->decomp_type = DecompositionTypes::SVD;
-    double energy = dmrg->solve(20, mps->center == 0, 1E-12);
+    long double energy = dmrg->solve(20, mps->center == 0, 1E-12);
 
     cout << "== " << name << " (DMRG) ==" << setw(20) << target
          << " E = " << fixed << setw(22) << setprecision(12) << energy
