+++ unit_test/test_fitpg_n2_631g.cpp
@@ -12,7 +12,7 @@ class TestFITPGN2631G : public ::testing::Test {
     typedef double FP;
 
     template <typename S, typename FL>
-    void test_dmrg(int n_ext, int ci_order, const S target, double energy,
+    void test_dmrg(int n_ext, int ci_order, const S target, long double energy,
                    const shared_ptr<HamiltonianQC<S, FL>> &hamil,
                    const shared_ptr<HamiltonianQC<S, FL>> &hamil_red,
                    const string &name, DecompositionTypes dt, NoiseTypes nt);
@@ -34,7 +34,7 @@ class TestFITPGN2631G : public ::testing::Test {
 
 template <typename S, typename FL>
 void TestFITPGN2631G::test_dmrg(
-    int n_ext, int ci_order, const S target, double energy,
+    int n_ext, int ci_order, const S target, long double energy,
     const shared_ptr<HamiltonianQC<S, FL>> &hamil,
     const shared_ptr<HamiltonianQC<S, FL>> &hamil_red, const string &name,
     DecompositionTypes dt, NoiseTypes nt) {
@@ -127,7 +127,7 @@ void TestFITPGN2631G::test_dmrg(
     dmrg->decomp_type = dt;
     dmrg->noise_type = nt;
     dmrg->decomp_last_site = dcl;
-    double ener = dmrg->solve(10, mps->center == 0, 1E-8);
+    long double ener = dmrg->solve(10, mps->center == 0, 1E-8);
 
     cout << "== " << name << " ==" << setw(20) << target << " E = " << fixed
          << setw(22) << setprecision(12) << ener << " error = " << scientific
@@ -238,7 +238,7 @@ void TestFITPGN2631G::test_dmrg(
     dmrg_red->decomp_type = dt;
     dmrg_red->noise_type = nt;
     dmrg_red->decomp_last_site = dcl;
-    double ener_red = dmrg_red->solve(20, mps_red->center == 0, 1E-8);
+    long double ener_red = dmrg_red->solve(20, mps_red->center == 0, 1E-8);
 
     cout << "== " << name << " ==" << setw(20) << target << " E = " << fixed
          << setw(22) << setprecision(12) << ener_red
@@ -265,7 +265,7 @@ TEST_F(TestFITPGN2631G, TestSU2) {
 
     SU2 vacuum(0);
     SU2 target(fcidump->n_elec(), 0, 0);
-    double energy = 0.1;
+    long double energy = 0.1;
 
     shared_ptr<HamiltonianQC<SU2, double>> hamil =
         make_shared<HamiltonianQC<SU2, double>>(vacuum, fcidump->n_sites(),
@@ -295,7 +295,7 @@ TEST_F(TestFITPGN2631G, TestSZ) {
 
     SZ vacuum(0);
     SZ target(fcidump->n_elec(), 0, 0);
-    double energy = 0.1;
+    long double energy = 0.1;
 
     shared_ptr<HamiltonianQC<SZ, double>> hamil =
         make_shared<HamiltonianQC<SZ, double>>(vacuum, fcidump->n_sites(),
